backend/ecommerce-backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose
alembic
fastapi-jwt-auth
passlib
pydantic_settings
python-multipart